Extract background index-scan subsystem from spec.c into index_scan.{c,h}
Pure refactor, no behavior change. Moves the keyspace scan/reindex machinery
out of spec.c into a dedicated translation unit:

  - scanner lifecycle (IndexesScanner_New/NewGlobal/Free/Cancel/ResetProgression)
  - background-indexing memory-limit / OOM helpers (now file-private)
  - per-key Indexes_ScanProc and the synchronous reindex task
  - IndexSpec_ScanAndReindex / Indexes_ScanAndReindex / reindex thread pool
  - the FT.DEBUG scanner (DebugIndexes*)

index_scan.h holds the public scan API plus the scanner / DebugIndexScannerCode
types (moved from spec.h); spec.h keeps only the forward declaration of
IndexesScanner and gains an IndexSpec_UpdateDoc prototype. The interleaved
non-scan code (setMemoryInfo and the shared memory-info layer, IndexSpecRef_*,
IndexSpec_DoneIndexingCallabck, Indexes_SetTempSpecsTimers) stays in spec.c.
spec.c, module.c, info_command.c and debug_commands.c include index_scan.h.
