+++ src/info/info_command.c
@@ -11,6 +11,7 @@
 
 #include "triemap_ffi.h"
 #include "spec.h"
+#include "index_scan.h"
 #include "inverted_index_ffi.h"
 #include "vector_index.h"
 #include "cursor.h"
