+++ src/module.c
@@ -28,6 +28,7 @@
 #include "rmutil/util.h"
 #include "rmutil/args.h"
 #include "spec.h"
+#include "index_scan.h"
 #include "util/logging.h"
 #include "util/workers.h"
 #include "util/references.h"
